upgraded refactor
