+++ screenmatch/src/main/java/com/jsrspring/screenmatch/model/RatingData.java
@@ -3,7 +3,7 @@ package com.jsrspring.screenmatch.model;
 import com.fasterxml.jackson.annotation.JsonAlias;
 
 // clasificación
-public record Rating(
+public record RatingData(
         @JsonAlias("Source") String source,
         @JsonAlias("Value") String value
 ) { }
