+++ screenmatch/src/main/java/com/jsrspring/screenmatch/model/Series.java
@@ -1,33 +1,119 @@
 package com.jsrspring.screenmatch.model;
 
-import com.fasterxml.jackson.annotation.JsonAlias;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.jsrspring.screenmatch.service.ChatGPTApiService;
+import com.theokanning.openai.OpenAiHttpException;
+import jakarta.persistence.*;
 
 import java.util.List;
+import java.util.OptionalDouble;
 
-@JsonIgnoreProperties(ignoreUnknown = true) // Ignora las propiedades que no sean incuidas en mi record
-public record Series(
-        @JsonAlias("Title") String title,
-        @JsonAlias("Year") String year,
-        @JsonAlias("Rated") String rated,
-        @JsonAlias("Released") String released,
-        @JsonAlias("Runtime") String runtime,
-        @JsonAlias("Genre") String genre,
-        @JsonAlias("Director") String director,
-        @JsonAlias("Writer") String writer,
-        @JsonAlias("Actors") String actors,
-        @JsonAlias("Plot") String plot,
-        @JsonAlias("Language") String language,
-        @JsonAlias("Country") String country,
-        @JsonAlias("Awards") String awards,
-        @JsonAlias("Poster") String poster,
-        @JsonAlias("Ratings") List<Rating> ratings,
-        @JsonAlias("Metascore") String metascore,
-        @JsonAlias("imdbRating") String imdbRating, // evaluation
-        @JsonAlias("imdbVotes") String imdbVotes,
-        @JsonAlias("imdbID") String imdbID,
-        @JsonAlias("Type") String type,
-        @JsonAlias("totalSeasons") Integer totalSeasons, // temporadas
-        @JsonAlias("Response") boolean response
-) {
+@Entity
+@Table(name = "series")
+public class Series {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(unique = true)
+    private String title;
+    private Integer totalSeasons;
+    private Double evaluation;
+    private String poster;
+    @Enumerated(EnumType.STRING)
+    private Category genre;
+    private String actors;
+    private String synopsis;
+    @Transient
+    private List<Episode> episodes;
+
+    public Series() {}
+
+    public Series(SeriesData seriesData) {
+        this.title = seriesData.title();
+        this.totalSeasons = seriesData.totalSeasons();
+        this.evaluation = OptionalDouble.of(Double.parseDouble(seriesData.imdbRating())).orElse(0);
+        this.poster = seriesData.poster();
+        this.genre = Category.fromString(seriesData.genre().split(",")[0].trim());
+        this.actors = seriesData.actors();
+        try {
+            this.synopsis = ChatGPTApiService.getTranslation(seriesData.plot());
+        } catch (OpenAiHttpException e) {
+            this.synopsis = seriesData.plot();
+            System.out.println("Error en Series " + e.getMessage());
+        }
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public Integer getTotalSeasons() {
+        return totalSeasons;
+    }
+
+    public void setTotalSeasons(Integer totalSeasons) {
+        this.totalSeasons = totalSeasons;
+    }
+
+    public Double getEvaluation() {
+        return evaluation;
+    }
+
+    public void setEvaluation(Double evaluation) {
+        this.evaluation = evaluation;
+    }
+
+    public String getPoster() {
+        return poster;
+    }
+
+    public void setPoster(String poster) {
+        this.poster = poster;
+    }
+
+    public Category getGenre() {
+        return genre;
+    }
+
+    public void setGenre(Category genre) {
+        this.genre = genre;
+    }
+
+    public String getActors() {
+        return actors;
+    }
+
+    public void setActors(String actors) {
+        this.actors = actors;
+    }
+
+    public String getSynopsis() {
+        return synopsis;
+    }
+
+    public void setSynopsis(String synopsis) {
+        this.synopsis = synopsis;
+    }
+
+    @Override
+    public String toString() {
+        return "title='" + title + '\'' +
+                ", totalSeasons=" + totalSeasons +
+                ", evaluation=" + evaluation +
+                ", poster='" + poster + '\'' +
+                ", genre=" + genre +
+                ", actors='" + actors + '\'' +
+                ", synopsis='" + synopsis;
+    }
 }
