+++ screenmatch/src/main/java/com/jsrspring/screenmatch/model/SeriesData.java
@@ -1,0 +1,33 @@
+package com.jsrspring.screenmatch.model;
+
+import com.fasterxml.jackson.annotation.JsonAlias;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import java.util.List;
+
+@JsonIgnoreProperties(ignoreUnknown = true) // Ignora las propiedades que no sean incuidas en mi record
+public record SeriesData(
+        @JsonAlias("Title") String title,
+        @JsonAlias("Year") String year,
+        @JsonAlias("Rated") String rated,
+        @JsonAlias("Released") String released,
+        @JsonAlias("Runtime") String runtime,
+        @JsonAlias("Genre") String genre,
+        @JsonAlias("Director") String director,
+        @JsonAlias("Writer") String writer,
+        @JsonAlias("Actors") String actors,
+        @JsonAlias("Plot") String plot,
+        @JsonAlias("Language") String language,
+        @JsonAlias("Country") String country,
+        @JsonAlias("Awards") String awards,
+        @JsonAlias("Poster") String poster,
+        @JsonAlias("Ratings") List<RatingData> ratingData,
+        @JsonAlias("Metascore") String metascore,
+        @JsonAlias("imdbRating") String imdbRating, // evaluation
+        @JsonAlias("imdbVotes") String imdbVotes,
+        @JsonAlias("imdbID") String imdbID,
+        @JsonAlias("Type") String type,
+        @JsonAlias("totalSeasons") Integer totalSeasons, // temporadas
+        @JsonAlias("Response") boolean response
+) {
+}
