+++ screenmatch/src/main/java/com/jsrspring/screenmatch/repository/SeriesRepository.java
@@ -1,6 +1,6 @@
 package com.jsrspring.screenmatch.repository;
 
-import com.jsrspring.screenmatch.model.SeriesDB;
+import com.jsrspring.screenmatch.model.Series;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeriesRepository extends JpaRepository<SeriesDB, Long> { }// nombre de la entidad y el tipo del identificador
+public interface SeriesRepository extends JpaRepository<Series, Long> { }// nombre de la entidad y el tipo del identificador
