+++ screenmatch/src/main/resources/application.properties
@@ -1,12 +1,9 @@
 spring.application.name=screenmatch
 
 # Configuracion en la DB
-#spring.datasource.url=jdbc:postgresql://${DB_HOST}/${DB_NAME}
-spring.datasource.url=jdbc:postgresql://localhost/series_db
-#spring.datasource.username=${DB_USER}
-spring.datasource.username=postgres
-#spring.datasource.password=${DB_PASSWORD}
-spring.datasource.password=123456
+spring.datasource.url=jdbc:postgresql://${DB_HOST}/${DB_NAME}
+spring.datasource.username=${DB_USER}
+spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=org.postgresql.Driver
 hibernate.dialect=org.hibernate.dialect.HSQLDialect
